refactor: #33 entity→domainに変更. 良案かは不明.
